refactor(#42): inventory 패키지 분리 및 InventoryController 추가
