+++ src/main/java/software_capstone/backend/app/inventory/repository/ShopItemRepository.java
@@ -1,4 +1,4 @@
-package software_capstone.backend.app.store.repository;
+package software_capstone.backend.app.inventory.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import software_capstone.backend.app.store.document.category.ItemCategory;
