+++ src/main/java/software_capstone/backend/app/store/service/ShopService.java
@@ -3,12 +3,13 @@ package software_capstone.backend.app.store.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import software_capstone.backend.app.inventory.service.InventoryService;
 import software_capstone.backend.app.store.document.category.ItemCategory;
 import software_capstone.backend.app.store.document.ShopItem;
 import software_capstone.backend.app.store.dto.request.PurchaseRequest;
 import software_capstone.backend.app.store.dto.response.PurchaseResponse;
 import software_capstone.backend.app.store.dto.response.ShopItemResponse;
-import software_capstone.backend.app.store.repository.ShopItemRepository;
+import software_capstone.backend.app.inventory.repository.ShopItemRepository;
 import software_capstone.backend.app.user.document.User;
 import software_capstone.backend.app.user.service.UserService;
 import software_capstone.backend.global.exception.ErrorMessage;
